requirements.txt
aiohttp==3.6.2
bs4==0.0.1
basc_py4chan==0.6.5
chemspipy==2.0.0
dateparser==1.0.0
discord.py==1.6.0
fastapi==0.63.0
feedparser==6.0.2
gtts==2.2.1
insprio==0.0.1
multidict==4.6.1
hentai==3.1.5
PixivPy==3.5.10
psycopg2-binary==2.8.6
Pybooru==4.2.0
pydrive2==1.4.10
PyNaCl==1.4.0
pysaucenao==1.5.0
python-youtube==0.6.4
requests==2.23.0
sauce-finder==2.2.4
selenium==3.141.0
udpy==2.0.0
urllib3==1.25.7
uvicorn[standard]==0.13.3
webdriver-manager==3.3.0
wordcloud==1.7.0
youtube-dl==2021.1.24.1
